dom/bindings/test/test_document_location_set_via_xray.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=905493
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 905493</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=905493">Mozilla Bug 905493</a>
<p id="display"></p>
<div id="content" style="display: none">
<iframe id="t" src="http://example.org/tests/dom/bindings/test/file_document_location_set_via_xray.html"></iframe>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 905493 **/

function test() {
  var doc = document.getElementById("t").contentWindow.document;
  ok(!("x" in doc), "Should have an Xray here");
  is(doc.x, undefined, "Really should have an Xray here");
  is(doc.wrappedJSObject.x, 5, "And wrapping the right thing");
  document.getElementById("t").onload = function() {
    ok(true, "Load happened");
    SimpleTest.finish();
  };
  try {
    // Test the forwarding location setter
    doc.location = "chrome://mochikit/content/tests/SimpleTest/test.css";
  } catch (e) {
    // Load failed
    ok(false, "Load failed");
    SimpleTest.finish();
  }
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(test);

</script>
</pre>
</body>
</html>
